% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ABBREV_AA}
\alias{ABBREV_AA}
\title{Amino acid abbreviation translations}
\format{
Named character vector defining single-letter character codes to 
         three-letter abbreviation mappings.
}
\usage{
ABBREV_AA
}
\description{
Mappings of amino acid abbreviations.
}
\examples{
aa <- c("Ala", "Ile", "Trp")
translateStrings(aa, ABBREV_AA)

}
\keyword{datasets}
