% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alakazam.R
\docType{package}
\name{alakazam-package}
\alias{alakazam-package}
\alias{_PACKAGE}
\title{alakazam: Immunoglobulin Clonal Lineage and Diversity Analysis}
\description{
Provides methods for high-throughput adaptive immune receptor repertoire sequencing (AIRR-Seq; Rep-Seq) analysis. In particular, immunoglobulin (Ig) sequence lineage reconstruction, lineage topology analysis, diversity profiling, amino acid property analysis and gene usage. Citations: Gupta and Vander Heiden, et al (2017) \doi{10.1093/bioinformatics/btv359}, Stern, Yaari and Vander Heiden, et al (2014) \doi{10.1126/scitranslmed.3008879}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alakazam.readthedocs.io/}
  \item Report bugs at \url{https://bitbucket.org/kleinstein/alakazam/issues}
}

}
\author{
\strong{Maintainer}: Susanna Marquez \email{susanna.marquez@yale.edu}

Authors:
\itemize{
  \item Namita Gupta \email{namita.gupta@yale.edu}
  \item Nima Nouri \email{nima.nouri@yale.edu}
  \item Ruoyi Jiang \email{ruoyi.jiang@yale.edu}
  \item Julian Zhou \email{julian.zhou@bulldogs.yale.edu}
  \item Kenneth Hoehn \email{kenneth.hoehn@yale.edu}
  \item Jason Vander Heiden \email{jason.vanderheiden@gmail.com}
  \item Steven Kleinstein \email{steven.kleinstein@yale.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item Daniel Gadala-Maria \email{daniel.gadala-maria@yale.edu} [contributor]
  \item Edel Aron \email{edel.aron@yale.edu} [contributor]
  \item Cole Jensen \email{cole.jensen@yale.edu} [contributor]
}

}
\keyword{internal}
