% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ChangeoClone-class}
\alias{ChangeoClone-class}
\alias{ChangeoClone}
\title{S4 class defining a clone}
\description{
\code{ChangeoClone} defines a common data structure for perform lineage recontruction
from Change-O data.
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame containing sequences and annotations. Contains the
columns \code{SEQUENCE_ID} and \code{SEQUENCE}, as well as any additional 
sequence-specific annotation columns.}

\item{\code{clone}}{string defining the clone identifier.}

\item{\code{germline}}{string containing the germline sequence for the clone.}

\item{\code{v_gene}}{string defining the V segment gene call.}

\item{\code{j_gene}}{string defining the J segment gene call.}

\item{\code{junc_len}}{numeric junction length (nucleotide count).}
}}

\seealso{
See \link{makeChangeoClone} and \link{buildPhylipLineage} for use.
}
