% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{cpuCount}
\alias{cpuCount}
\title{Available CPU cores}
\usage{
cpuCount()
}
\value{
Count of available cores. Returns 1 if undeterminable.
}
\description{
\code{cpuCount} determines the number of CPU cores available.
}
\examples{
cpuCount()

}
