% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lineage.R
\docType{data}
\name{ExampleTrees}
\alias{ExampleTrees}
\title{Example Ig lineage trees}
\format{A list of igraph objects output by \code{\link{buildPhylipLineage}}.}
\usage{
ExampleTrees
}
\description{
A set of Ig lineage trees generated from the \code{extdata/ExampleDb.gz} file.
}
\keyword{datasets}

