% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{gridPlot}
\alias{gridPlot}
\title{Plot multiple ggplot objects}
\usage{
gridPlot(..., ncol = 1)
}
\arguments{
\item{...}{ggplot objects to plot.}

\item{ncol}{number of columns in the plot.}
}
\description{
Plots multiple ggplot objects in an equally sized grid.
}
\references{
Modified from:
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)
}
\seealso{
\link{ggplot}.
}
