% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{IMGT_REGIONS}
\alias{IMGT_REGIONS}
\title{IMGT V-segment regions}
\format{
A list with regions named one of \code{c("fwr1", "cdr1", "fwr2", "cdr2", "fwr3")} 
         with values containing a numeric vector of length two defining the 
         \code{c(start, end)} positions of the named region.
}
\usage{
IMGT_REGIONS
}
\description{
A list defining the boundaries of V-segment framework regions (FWRs) and complementarity 
determining regions (CDRs) for IMGT-gapped immunoglobulin (Ig) nucleotide sequences 
according to the IMGT numbering scheme.
}
\references{
\url{http://imgt.org}
}
\keyword{datasets}
