% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{IUPAC_CODES}
\alias{IUPAC_AA}
\alias{IUPAC_CODES}
\alias{IUPAC_DNA}
\title{IUPAC ambiguous characters}
\format{A list with single character codes as names and values containing character 
         vectors that define the set of standard characters that match to each each 
         ambiguous character.
\itemize{
  \item  \code{IUPAC_DNA}:  DNA ambiguous character translations.
  \item  \code{IUPAC_AA}:   Amino acid ambiguous character translations.
}}
\usage{
IUPAC_DNA

IUPAC_AA
}
\description{
A translation list mapping IUPAC ambiguous characters code to corresponding nucleotide
amino acid characters.
}
\keyword{datasets}

