% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{makeTempDir}
\alias{makeTempDir}
\title{Create a temporary folder}
\usage{
makeTempDir(prefix)
}
\arguments{
\item{prefix}{prefix name for the folder.}
}
\value{
The path to the temporary folder.
}
\description{
\code{makeTempDir} creates a randomly named temporary folder in the 
system temp location.
}
\examples{
makeTempDir("Clone50")

}
\seealso{
This is just a wrapper for \code{\link{tempfile}} and 
          \code{\link{dir.create}}.

Other file input and output functions: \code{\link{readChangeoDb}},
  \code{\link{writeChangeoDb}}
}

