% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{progressBar}
\alias{progressBar}
\title{Standard progress bar}
\usage{
progressBar(n)
}
\arguments{
\item{n}{maximum number of ticks}
}
\value{
A \link[progress]{progress_bar} object.
}
\description{
\code{progressBar} defines a common progress bar format.
}
