% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{SingleDb}
\alias{SingleDb}
\title{Single sequence AIRR database}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 32 columns.
}
\usage{
SingleDb
}
\description{
A database with just one sequence from \code{ExampleDb} and additional AIRR Rearrangement fields 
containing alignment information. The sequence was reanalyzed with a recent versions of 
alignment software (IgBLAST 1.16.0) and reference germlines (IMGT 2020-08-12).
}
\seealso{
\link{ExampleDb}
}
\keyword{datasets}
