% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{writeChangeoDb}
\alias{writeChangeoDb}
\title{Write a Change-O tab-delimited database file}
\usage{
writeChangeoDb(data, file)
}
\arguments{
\item{data}{data.frame of Change-O data.}

\item{file}{output file name.}
}
\description{
\code{writeChangeoDb} is a simple wrapper around \code{\link{write.table}} with defaults 
appropriate for writing a Change-O tab-delimited database file from a data.frame.
}
\examples{
\dontrun{
  # Write a database
  writeChangeoDb(data, "changeo_output.tab")
}

}
\seealso{
Wraps \code{\link{write.table}}.

Other file input and output functions: \code{\link{makeTempDir}},
  \code{\link{readChangeoDb}}
}

