% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ExampleTrees}
\alias{ExampleTrees}
\title{Example Ig lineage trees}
\format{
A list of igraph objects output by \link{buildPhylipLineage}.
          Each node of each tree has the following annotations (vertex attributes):
  \itemize{
    \item  \code{sample_id}:          Sample identifier(s). Time in relation to vaccination.
    \item  \code{c_call}:             Isotype assignment(s). 
    \item  \code{duplication_count}:  Copy count (number of duplicates) of the sequence.
  }
}
\usage{
ExampleTrees
}
\description{
A set of Ig lineage trees generated from the \code{ExampleDb} file, subset to
only those trees with at least four nodes.
}
\seealso{
\link{ExampleTrees}
}
\keyword{datasets}
