% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diversity.R
\name{plotDiversityCurve}
\alias{plotDiversityCurve}
\title{Plot the results of alphaDiversity}
\usage{
plotDiversityCurve(
  data,
  colors = NULL,
  main_title = "Diversity",
  legend_title = "Group",
  log_x = FALSE,
  log_y = FALSE,
  xlim = NULL,
  ylim = NULL,
  annotate = c("none", "depth"),
  score = c("diversity", "evenness"),
  silent = FALSE,
  ...
)
}
\arguments{
\item{data}{\link{DiversityCurve} object returned by 
\link{alphaDiversity}.}

\item{colors}{named character vector whose names are values in the 
\code{group} column of the \code{data} slot of \code{data},
and whose values are colors to assign to those group names.}

\item{main_title}{string specifying the plot title.}

\item{legend_title}{string specifying the legend title.}

\item{log_x}{if \code{TRUE} then plot \eqn{q} on a log scale;
if \code{FALSE} plot on a linear scale.}

\item{log_y}{if \code{TRUE} then plot the diversity/evenness scores 
on a log scale; if \code{FALSE} plot on a linear scale.}

\item{xlim}{numeric vector of two values specifying the 
\code{c(lower, upper)} x-axis limits.}

\item{ylim}{numeric vector of two values specifying the 
\code{c(lower, upper)} y-axis limits.}

\item{annotate}{string defining whether to added values to the group labels 
of the legend. When \code{"none"} (default) is specified no
annotations are added. Specifying (\code{"depth"}) adds 
sequence counts to the labels.}

\item{score}{one of \code{"diversity"} or \code{"evenness"} specifying which
score to plot on the y-asis.}

\item{silent}{if \code{TRUE} do not draw the plot and just return the ggplot2 
object; if \code{FALSE} draw the plot.}

\item{...}{additional arguments to pass to ggplot2::theme.}
}
\value{
A \code{ggplot} object defining the plot.
}
\description{
\code{plotDiversityCurve} plots a \code{DiversityCurve} object.
}
\examples{
# Calculate diversity
div <- alphaDiversity(ExampleDb, group="sample_id", nboot=100)

# Plot diversity
plotDiversityCurve(div, legend_title="Sample")

#' # Plot diversity
plotDiversityCurve(div, legend_title="Sample", score="evenness")

}
\seealso{
See \link{alphaDiversity} and \link{alphaDiversity} for generating 
          \link{DiversityCurve} objects for input. Plotting is performed with \link{ggplot}.
}
